+++ khachhang.h
@@ -18,21 +18,9 @@ public:
     static void DeleteKhachHang(vector<KhachHang>&);
     static void EditKhachHang(vector<KhachHang>&);
     static void ShowKhachHang(vector<KhachHang>&);
-    static void Show1KhachHang(vector<KhachHang>&, string);
+    void Show1KhachHang(vector<KhachHang>&);
     static int CheckTonTai(vector<KhachHang>&, string);
     void WriteToFileKhachHang(string);
-    void setMaKhachHang(string);
-    void setTenKhachHang(string);
-    void setDiaChi(string);
-    void setDienThoai(string);
-    void setNamSinh(int);
-    void setGioiTinh(string);
-    string getMaKhachHang();
-    string getTenKhachHang();
-    string getDiaChi();
-    string getDienThoai();
-    int getNamSinh();
-    string getGioiTinh();   
     friend class HoaDon;
 };
 #endif